+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.3.0] - 2020-01-27
+### Added
+ - The `add_unit_amenity_property()` method, which uses [v1/unit-amenity-properties](https://connect.vacasa.com/#operation/create-unit-amenity-properties)
+ - The `update_unit_amenity_property()` method, which uses [/v1/unit-amenity-properties/{id}](https://connect.vacasa.com/#operation/update-unit-amenity-properties)
+
 ## [4.2.0] - 2020-01-27
 ### Added
  - The `get_unit_amenity_properties()` method, which uses [v1/unit-amenity-properties](https://connect.vacasa.com/#operation/get-unit-amenity-properties)